Rename demo package to rrbot_control_boilerplate
There is already an `rrbot_control` package from the
`gazebo_ros_demos` repo[1].  Give this package a unique name to
resolve conflicts.

[1]: https://github.com/ros-simulation/gazebo_ros_demos
