+++ rrbot_control_boilerplate/launch/rrbot_simulation.launch
@@ -16,7 +16,9 @@
           output="screen" launch-prefix="$(arg launch_prefix)"/>
 
     <!-- Load controller settings -->
-    <rosparam file="$(find ros_control_boilerplate)/rrbot_control/config/rrbot_controllers.yaml" command="load"/>
+    <rosparam
+        file="$(find rrbot_control_boilerplate)/config/rrbot_controllers.yaml"
+        command="load"/>
 
     <!-- Load controller manager -->
     <node name="ros_control_controller_manager" pkg="controller_manager" type="controller_manager" respawn="false"
