+++ rrbot_control_boilerplate/launch/rrbot_test_trajectory.launch
@@ -12,7 +12,9 @@
     <node name="test_trajectory" pkg="ros_control_boilerplate" type="test_trajectory"
           output="screen" launch-prefix="$(arg launch_prefix)">
       <param name="action_topic" value="/rrbot/position_trajectory_controller/follow_joint_trajectory/"/>
-      <rosparam file="$(find ros_control_boilerplate)/rrbot_control/config/rrbot_controllers.yaml" command="load"/>
+      <rosparam
+          file="$(find rrbot_control_boilerplate)/config/rrbot_controllers.yaml"
+          command="load"/>
     </node>
 
   </group>
